refactor: move registration fragments to ui/registration package
