+++ app/src/main/java/br/com/connectattoo/ui/registration/UserRegistration.kt
@@ -1,4 +1,4 @@
-package br.com.connectattoo.auth
+package br.com.connectattoo.ui.registration
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import android.widget.EditText
 import androidx.fragment.app.Fragment
 import androidx.viewbinding.ViewBinding
-import br.com.connectattoo.R
 
 abstract class UserRegistration<T: ViewBinding> : Fragment() {
