+++ app/src/main/res/layout/fragment_tattoo_client_registration.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fillViewport="true"
     android:background="@color/white"
-    tools:context=".auth.userRegistration.TattooClientRegistrationFragment">
+    tools:context=".ui.registration.TattooClientRegistrationFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
